mctp-echo, mctp-req: support variable-sized payloads
This change allows mctp-echo and mctp-req to send and receive messages
of greater than one byte.

mctp-echo will simply echo the same packet it received, whatever the
size.

mctp-req now takes a 'len' argument, which specifies the buffer size to
use. This buffer is filled with sequential bytes, and compared on
receive.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>