mctpd: Add bus owner and endpoint switch using arguments
Currently, mctpd defaults to run as a bus owner, so ctx->bus_owner is
always true. This commit adds a role argument so that we can change the
role without editing the source code.

Tested: Send a Get EID command to mctpd. See the response:

- Running as Bus Owner (no flag):

      00 00 02 00 ee 12 00

- Running as Endpoint (-r endpoint):

      00 00 02 00 ee 02 00

The second to last byte signifies the endpoint type.
(0xee is the EID assigned to mctpd)

Signed-off-by: Khang Nguyen <khangng@amperecomputing.com>