tests: introduce MCTP infrastructure test facility
Incroduce some pytest-based infrastructure for MCTP components. This is
currently targeted at `mctpd`, but would also work for `mctp` testing
too.

Most of the change here is in a pytest fixture to wrap a MCTP utility:
we use the new ops abstractions to provide mock implementations of both a
netlink interface, and AF_MCTP socket support.

We introduce a single test as a baseline: test_mctpd.py, which just
calls SetupEndpoint and expects routing/neighbour table updates.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>