mctpd: Cope with devices that don't support `Get Endpoint UUID`
It turns out that we don't allocate peer->uuid in the initial peer setup
path if `Get Endpoint UUID` fails:

https://github.com/CodeConstruct/mctp/blob/858d20362329069ec29dc39011f8b960ed56f8ba/src/mctpd.c#L1918-L1923

That behaviour invalidates the assumption that peer->uuid was always
valid if the peer was valid. For the case where peer->uuid is not valid,
take the path of treating it as an unrecognised device.

Fixes: 7ec2f8daa3a8 ("mctpd: Add support for endpoint recovery")
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>