mctpd: Fix resource handling over recovery of exchanged peer devices
A fundamental flaw merged in 7ec2f8daa3a8 ("mctpd: Add support for
endpoint recovery") was that NULL was passed for the peer parameter in
the call to endpoint_assign_eid(). I had intended to circle back to
that, but it clearly got overlooked in testing and review.

While this caused an obvious crash when the path was hit, the subsequent
resource cleanup was also in error: As the original peer has been
removed there's no need for the remaining state maintenance executed in
the original flow. Instead, immediately return with the result of
endpoint_assign_eid() against the new peer device.

Fixes: 7ec2f8daa3a8 ("mctpd: Add support for endpoint recovery")
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>