meson: Change `test` option type from `feature` to `boolean`
The user needs to choose whether tests are enabled or not while
configuring the build, or fallback to the provided default. We can't
divine their thoughts; there's no programmatic way to determine a valid
choice, unlike with dependency-detection via pkg-config or some other
mechanism.

The `required` parameter of `find_program()` defaults to true. Eliminate
its specification for test programs, as the `find_program()` invocations
are already gated on the result of `get_option('tests')`.

Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>