mctp-netlink: Add a facility to store external data on links
For upcoming mctpd changes, we will need a facility for storing
dbus-specific information on each link. This change adds a

  void * userdata

pointer to the links, and accessors for callers to get/set this pointer.
For convenience, we also pass this pointer on change events, allowing
for a quick way to release/update the userdata on link delete/change.

WIP: untested.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>