mctp-bench: Add a utility tool to send/receive messages and display throughput
A utility tool that implements a sender and receiver to benchmark mctp
throughput. This is split into a sender command and a receive command in the
same executable.

The receive command waits for incoming messages, notifies the user of first
message received and then prints out measured throughput periodically.

Example usage:
	mctp-bench recv

The sender command sends messages continously, the size of which can be
set by the user, defaulting to 8 bytes.

Example usage:
	mctp-bench send len 1000 eid 8

Signed-off-by: Timon Liu <timon@codeconstruct.com.au>