tests: notify netlink socket on system link/neighbour/address/route changes
If we update the system network configuration (by adding/removing links,
neighbours, addresses or routes) after mctpd has performed the nl dump
operations, mctpd will not be aware of those changes.

Add notifiers to the {add,del}_{link,address,neighbour,route}
functions on System, which send an appropriate RTM_ message informing
mctpd of the change. This requires the system to have a reference to the
NLSocket, and for those functions to become async.

Since we're now creating netlink messages for the
address/route/link/neighbour in multiple sites, use internal _format
helpers for the rtnetlink attribute construction.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>