mctpd: use object vtable for endpoints
Currently, we use a set of fallback vtable for our entire dbus subtree,
and detect the correct objects in endpoint-specific callbacks, and do
a peer lookup (from the dbus path) to find the struct peer.

Instead, establish a bus object for endpoints, storing the struct peer
in the dbus userdata. vtable methods can then just extract the peer
directly.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>