mctpd: use object vtable for links
Currently, we use a set of fallback vtable for our entire dbus subtree,
and detect the correct objects in endpoint-specific callbacks, and do
a link lookup (from the dbus path) to find the struct peer.

Instead, establish a bus object for links, storing the struct link
in the dbus userdata. vtable methods can then just extract the link
directly.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>