mctpd: de-duplicate logic in enumerate_objects()
Currently, enumerate_objects() repeats the enumeration logic: once to
count the number of paths required, and a second to populate the paths.

Rather than duplicate the code, allow calling enumerate_objects() with a
null paths array, for the initial counting. We will only populate the
paths array when non-null (and the caller has allocated
appropriately-sized storage).

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>