tests: separate test fixtures into a library, and add a utility main()
The pytest fixtures in conftest.py contain the entire fake mctp network
and netlink interfaces. While these are handy for the tests, it would
also be nice to be able to run test-mctpd binary under the fake
environment, not under pytest.

Separate the wrapper code into a python module 'mctpd', which conftest
imports for the test fixtures. Add a small main which runs the fake
mctpd environment using the default sysnet configuration.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>