mctpd: Add Network1.LearnEndpoint
Our current support for bridged endpoints requires static EIDs to be
allocated in advance. To handle that case, add a new dbus call:
LearnEndpoint, on the net object (au.com.codeconstruct.MCTP.Network1).

This takes an EID, performs a discovery query, and publishes a new
endpoint object if that query was sucessful. This allows dbus clients to
enumerate briged EPs.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>