mctpd: Allocate peers as separate (stable) objects
Currently, we allocate peers in an array, stored as ctx->peers. We
realloc this array on demand, so the peer pointers are not stable.

We would like to use peers as proper dbus objects, which will require
stable pointers. Instead of the array, allocate the peers individually,
and convert the ctx->peers array to an array-of-pointers, which may be
realloced freely.

Since the pointers array is small, just realloc individually instead of
batching.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>