tests: disable TAP in interactive terminal
pytest-tap output is great for just summaries, but the log is very much
unreadable if tests fail. Currently there is an unresolved issue related
to terminalreporter, see [1].

[1] https://github.com/python-tap/pytest-tap/issues/98

This commit adds a way to disable TAP. Newer versions of meson support
an --interactive flag, like so:

    meson test -C obj --interactive

This makes the test run directly in the terminal, not redirected to
a file. When this happens, i.e stdout is a terminal, we run the test
without TAP.

Signed-off-by: Khang D Nguyen <khangng@os.amperecomputing.com>