mctpd: improve handling for out-of-spec Set Endpoint ID responses
We may have endpoints reporting strange EID values in their set endpoint
ID response, even when reporting that the assignement was accepted.
Particularly, if the EID is in the invalid range, we will still attempt
to add neighour and route entries for this invalid EID.

We still want to accommodate valid IDs that were not the same as the one
assigned, but ensure the EID is valid before processing.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>