mctpd: allow bridge allocation with SetupEndpoint
Currently, we only allow bridge pool allocation through an
AssignEndpoint call, as that is guaranteed to involve a Set Endpoint ID
command, required to start the pool allocation process.

LearnEndpoint is intended to never modify endpoint state, so we keep that
as-is.

SetupEndpoint has always been a convenience method, intended to do a
LearnEndpoint if possible, or fall back to AssignEndpoint if not.
Because of this, we are not making any assurances about preserving state
with SetupEndpoint.

With the new bridge support, the current distinction between
AssignEndpoint (which will allocate a bridge pool) and SetupEndpoint
(which will not) is a potential point of confusion. Instead, allow
bridge allocations through SetupEndpoint.

We have a new conditional path here: we try the LearnEndpoint (ie, a Get
Endpoint ID, to see if we can use that EID) first, but add a new check
to determine if this is a bridge EID type. Is so, we force the fallback
to Set Endpoint ID, which will allow a bridge EID allocation too.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>