mctpd: prevent add-then-remove during SetupEndpoint reassignment
Currently, if an endpoint needs reassigning due to a bridge conflict,
SetupEndpoint may publish an endpoint, then remove it. This is because
we're calling get_endpoint_peer() for the Get Endpoint ID function
(which adds the peer), but later check for bridge compatibility.

Prevent this by open-coding the parts we need from get_endpoint_peer,
and only performing the add once we have a valid peer. This requires a
bit of re-work that is particular to bridge allocation in the
SetupEndpoint case.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>