mctpd: allow for pool allocations that are smaller than our max limit
Currently, we attempt to allocate bridge pool ranges of the size of our
max pool configuration setting, and then trim after we know the
requested pool size. If the max allocation is not available, we do not
provide *any* EID range to the requesting bridge.

However, it's entirely likely that the bridge will request a pool that
is smaller than our maximum. We should not reject that allocation, as
there is space available.

Instead of insisting on allocating the max, just pre-allocate the
largest space up to the max. When we then learn the bridge pool size,
offer the allocation that we made. If this is smaller than the
preallocation, we trim. If it is larger, we just offer what is
allocated.

This changes a failure case, where the tests expect the less-than-max
allocation to fail. No need to preserve this behaviour, as we can
actually offer a workable pool at this point.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>